+++ index.js
@@ -1,0 +1,13 @@
+module.exports = {
+    'clause': require('./dist_node/clause'),
+    'declaration': require('./dist_node/declaration'),
+    'expression': require('./dist_node/expression'),
+    'node': require('./dist_node/node'),
+    'package': require('./dist_node/package')
+    'pattern': require('./dist_node/pattern'),
+    'program': require('./dist_node/program'),
+    'serialization': require('./dist_node/serialization'),
+    'statement': require('./dist_node/statement'),
+    'token': require('./dist_node/token'),
+    'value': require('./dist_node/value')
+};